Extract body class because of findbugs
